Update SvmParams type

+++ src/accounts/worker/svm_params.rs
@@ -1,9 +1,9 @@
 #[repr(C)]
 #[derive(Copy, Clone)]
 pub struct SvmParams {
-    pub v: i64,
-    pub psi: i64,
-    pub p: i64,
-    pub c: i64,
-    pub vt: i64,
-} // size: 40 bytes
+    pub v: f64,
+    pub psi: f64,
+    pub p: f64,
+    pub c: f64,
+    pub vt: f64,
+} // s
